Moved logging to separate class which gets injected into lexer.
